+++ modules/safe/scripts/deploy-create2.ts
@@ -20,6 +20,7 @@ import CompatibilityFallbackHandler from "@safe-global/safe-smart-account/build/
 import MultiSend from "@safe-global/safe-smart-account/build/artifacts/contracts/libraries/MultiSend.sol/MultiSend.json";
 import MultiSendCallOnly from "@safe-global/safe-smart-account/build/artifacts/contracts/libraries/MultiSendCallOnly.sol/MultiSendCallOnly.json";
 import SafeMigration from "@safe-global/safe-smart-account/build/artifacts/contracts/libraries/SafeMigration.sol/SafeMigration.json";
+import SafeToL2Setup from "@safe-global/safe-smart-account/build/artifacts/contracts/libraries/SafeToL2Setup.sol/SafeToL2Setup.json";
 import SafeProxyFactory from "@safe-global/safe-smart-account/build/artifacts/contracts/proxies/SafeProxyFactory.sol/SafeProxyFactory.json";
 import Safe from "@safe-global/safe-smart-account/build/artifacts/contracts/Safe.sol/Safe.json";
 import SafeL2 from "@safe-global/safe-smart-account/build/artifacts/contracts/SafeL2.sol/SafeL2.json";
@@ -55,6 +56,7 @@ const CONTRACTS: { name: string; bytecode: Hex }[] = [
   { name: "CompatibilityFallbackHandler", bytecode: fallbackHandler },
   { name: "MultiSend", bytecode: MultiSend.bytecode as Hex },
   { name: "MultiSendCallOnly", bytecode: MultiSendCallOnly.bytecode as Hex },
+  { name: "SafeToL2Setup", bytecode: SafeToL2Setup.bytecode as Hex },
   {
     name: "SafeMigration",
     bytecode: concatHex([